Added Keycloak OIDC authentication
Added optional Keycloak OIDC authentication with example docker-compose.
Note that the catalog service has been moved to port 8082 in the
docker-compose so that the Keycloak server is available at the same
address both on the host machine and in the docker containers. This is
because the Keycloak Docker container was not playing well with
different port numbers. In a real deployment, the catalog service can
have port 8080 and Keycloak would be made accessible via the hostname of
the server.

Also updated Dockerfiles to pull from Docker hub rather than
code.ornl.gov.

Addresses issues #1 #2 and #3